change the projectname
