+++ AntDesign.Avalonia.sln
@@ -32,8 +32,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AntDesign.Sample.Web", "Sou
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avalonia.SourceGenerator", "Source\Avalonia.SourceGenerator\Avalonia.SourceGenerator.csproj", "{8148DE74-5328-4198-8067-8B6339B9CAB8}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avalonia.Toolkit", "Source\Avalonia.Toolkit\Avalonia.Toolkit.csproj", "{4832D4A7-B65C-4AE7-8997-2DCE7E8410DF}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avalonia.ReactiveUI.Toolkit", "Source\Avalonia.ReactiveUI.Toolkit\Avalonia.ReactiveUI.Toolkit.csproj", "{5CE934BE-6C44-4893-B6EC-F760AFA448BA}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AntDesign.Icons", "Source\AntDesign.Icons\AntDesign.Icons.csproj", "{CEE81A97-90B0-4959-AAE4-4E78771AE380}"
@@ -51,6 +49,8 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".Packages", ".Packages", "{3C61C340-656F-4BF8-8C2C-E5DB51BB0866}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AntDesign.Avalonia", "Packages\AntDesign.Avalonia\AntDesign.Avalonia.csproj", "{084349FA-3FCE-43CE-90C2-AF382348929C}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AntDesign.Toolkit", "Source\AntDesign.Toolkit\AntDesign.Toolkit.csproj", "{744B9A51-3DE4-4631-AA8F-414455520EC4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -90,10 +90,6 @@ Global
 		{8148DE74-5328-4198-8067-8B6339B9CAB8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8148DE74-5328-4198-8067-8B6339B9CAB8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8148DE74-5328-4198-8067-8B6339B9CAB8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4832D4A7-B65C-4AE7-8997-2DCE7E8410DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4832D4A7-B65C-4AE7-8997-2DCE7E8410DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4832D4A7-B65C-4AE7-8997-2DCE7E8410DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4832D4A7-B65C-4AE7-8997-2DCE7E8410DF}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5CE934BE-6C44-4893-B6EC-F760AFA448BA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5CE934BE-6C44-4893-B6EC-F760AFA448BA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5CE934BE-6C44-4893-B6EC-F760AFA448BA}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -122,6 +118,10 @@ Global
 		{084349FA-3FCE-43CE-90C2-AF382348929C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{084349FA-3FCE-43CE-90C2-AF382348929C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{084349FA-3FCE-43CE-90C2-AF382348929C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{744B9A51-3DE4-4631-AA8F-414455520EC4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{744B9A51-3DE4-4631-AA8F-414455520EC4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{744B9A51-3DE4-4631-AA8F-414455520EC4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{744B9A51-3DE4-4631-AA8F-414455520EC4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
