+++ Source/AntDesign.Sample.Browser.Blazor/Program.cs
@@ -1,7 +1,7 @@
 ﻿using Avalonia;
 using Avalonia.Browser.Blazor;
 using Avalonia.ReactiveUI;
-using Avalonia.Toolkit;
+using AntDesign.Toolkit;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 using System.Runtime.Versioning;
 using antdesignApp = AntDesign.Sample.App;
@@ -21,7 +21,7 @@ public class Program
     {
         await AppBuilder.Configure<antdesignApp>()
                         .UseReactiveUI()
-                        .UseAvaloniaToolkit()
+                        .UseAntDesignToolkit()
                         .StartBlazorAppAsync();
     }
