+++ Source/AntDesign.Sample.Desktop/Program.cs
@@ -1,6 +1,6 @@
 ﻿using Avalonia;
 using Avalonia.Controls;
-using Avalonia.Toolkit;
+using AntDesign.Toolkit;
 using Avalonia.ReactiveUI;
 
 namespace AntDesign.Sample;
@@ -18,7 +18,7 @@ internal class Program
     public static AppBuilder BuildAvaloniaApp()
         => AppBuilder.Configure<App>()
                      .UsePlatformDetect()
-                     .UseAvaloniaToolkit()
+                     .UseAntDesignToolkit()
                      .LogToTrace()
                      .UseReactiveUI();
 }
