+++ Source/AntDesign.Sample.Shared/Views/DialogsView.axaml.cs
@@ -1,7 +1,7 @@
 using AntDesign.Sample.ViewModels;
 using Avalonia.Dialogs;
 using Avalonia.Platform.Storage;
-using Avalonia.Toolkit.Helpers;
+using AntDesign.Toolkit.Helpers;
 
 namespace AntDesign.Sample.Views;
 public partial class DialogsView : ReactiveUserControl<DialogsViewModel>
