+++ Source/AntDesign.Sample.Web/Program.cs
@@ -2,7 +2,7 @@
 using Avalonia;
 using Avalonia.Browser;
 using Avalonia.ReactiveUI;
-using Avalonia.Toolkit;
+using AntDesign.Toolkit;
 using System.Runtime.Versioning;
 
 [assembly: SupportedOSPlatform("browser")]
@@ -13,7 +13,7 @@ internal partial class Program
     {
        await BuildAvaloniaApp()
             .UseReactiveUI()
-            .UseAvaloniaToolkit()
+            .UseAntDesignToolkit()
             .StartBrowserAppAsync("out");
     }
