+++ Source/AntDesign.Toolkit/Media/AntDesignFontCollection.cs
@@ -1,4 +1,4 @@
-﻿namespace Avalonia.Toolkit.Media;
+﻿namespace AntDesign.Toolkit.Media;
 internal sealed class AntDesignFontCollection : EmbeddedFontCollection
 {
     public AntDesignFontCollection(Uri key, Uri source) : base(key, source)
