+++ Source/AntDesign.Toolkit/Properties/AssemblyInfo.cs
@@ -1,0 +1,12 @@
+﻿using Avalonia.Metadata;
+
+[assembly: XmlnsPrefix("https://github.com/avaloniaui", $"{nameof(AntDesign)}.{nameof(AntDesign.Toolkit)}")]
+[assembly: XmlnsDefinition("https://github.com/avaloniaui", $"{nameof(AntDesign)}.{nameof(AntDesign.Toolkit)}")]
+[assembly: XmlnsDefinition("https://github.com/avaloniaui", $"{nameof(AntDesign)}.{nameof(AntDesign.Toolkit)}.{nameof(AntDesign.Toolkit.Media)}")]
+[assembly: XmlnsDefinition("https://github.com/avaloniaui", $"{nameof(AntDesign)}.{nameof(AntDesign.Toolkit)}.{nameof(AntDesign.Toolkit.Settings)}")]
+ 
+
+[assembly: XmlnsPrefix("https://ant.design.com/avalonia", $"{nameof(AntDesign)}.{nameof(AntDesign.Toolkit)}")]
+[assembly: XmlnsDefinition("https://ant.design.com/avalonia", $"{nameof(AntDesign)}.{nameof(AntDesign.Toolkit)}")]
+[assembly: XmlnsDefinition("https://ant.design.com/avalonia", $"{nameof(AntDesign)}.{nameof(AntDesign.Toolkit)}.{nameof(AntDesign.Toolkit.Media)}")]
+[assembly: XmlnsDefinition("https://ant.design.com/avalonia", $"{nameof(AntDesign)}.{nameof(AntDesign.Toolkit)}.{nameof(AntDesign.Toolkit.Settings)}")]
