--- Source/Avalonia.Toolkit/Properties/AssemblyInfo.cs
@@ -1,12 +1,0 @@
-﻿using Avalonia.Metadata;
-
-[assembly: XmlnsPrefix("https://github.com/avaloniaui", $"{nameof(Avalonia)}.{nameof(Avalonia.Toolkit)}")]
-[assembly: XmlnsDefinition("https://github.com/avaloniaui", $"{nameof(Avalonia)}.{nameof(Avalonia.Toolkit)}")]
-[assembly: XmlnsDefinition("https://github.com/avaloniaui", $"{nameof(Avalonia)}.{nameof(Avalonia.Toolkit)}.{nameof(Avalonia.Toolkit.Media)}")]
-[assembly: XmlnsDefinition("https://github.com/avaloniaui", $"{nameof(Avalonia)}.{nameof(Avalonia.Toolkit)}.{nameof(Avalonia.Toolkit.Settings)}")]
- 
-
-[assembly: XmlnsPrefix("https://ant.design.com/avalonia", $"{nameof(Avalonia)}.{nameof(Avalonia.Toolkit)}")]
-[assembly: XmlnsDefinition("https://ant.design.com/avalonia", $"{nameof(Avalonia)}.{nameof(Avalonia.Toolkit)}")]
-[assembly: XmlnsDefinition("https://ant.design.com/avalonia", $"{nameof(Avalonia)}.{nameof(Avalonia.Toolkit)}.{nameof(Avalonia.Toolkit.Media)}")]
-[assembly: XmlnsDefinition("https://ant.design.com/avalonia", $"{nameof(Avalonia)}.{nameof(Avalonia.Toolkit)}.{nameof(Avalonia.Toolkit.Settings)}")]
