--- Source/Avalonia.Toolkit/Properties/_globalusing.cs
@@ -1,4 +1,0 @@
-global using Avalonia.Media;
-global using Avalonia.Media.Fonts;
-global using System.Diagnostics.CodeAnalysis;
-global using System.Runtime.InteropServices;
